Revert "common/sclang: more integer type fixes"
This reverts commit d5275094655f8016cc2d205b7ad45317ef4a9f27.

+++ common/SC_AllocPool.cpp
@@ -535,7 +535,7 @@ void* AllocPool::Realloc(void* inPtr, size_t inReqSize)
 void AllocPool::LinkFree(AllocChunkPtr inChunk)
 {
 	size_t size = inChunk->Size();
-	size_t index = BinIndex(size);
+	int index = BinIndex(size);
 
 	AllocChunkPtr bin = mBins + index;
